# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/common/code/verify": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "send verify code",
                "parameters": [
                    {
                        "description": "email",
                        "name": "email",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "phone",
                        "name": "phone",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.ResponseSuccessBean-NullJson"
                        }
                    }
                }
            }
        },
        "/api/v1/customer/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "customer register",
                "parameters": [
                    {
                        "description": "customer register request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CustomerRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.ResponseSuccessBean-dto_CustomerRegisterResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CustomerRegisterReq": {
            "type": "object",
            "required": [
                "code",
                "password",
                "phone"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "dto.CustomerRegisterResp": {
            "type": "object",
            "properties": {
                "loginToken": {
                    "type": "string"
                }
            }
        },
        "result.NullJson": {
            "type": "object"
        },
        "result.ResponseSuccessBean-NullJson": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/result.NullJson"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "result.ResponseSuccessBean-dto_CustomerRegisterResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/dto.CustomerRegisterResp"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}